Break out parts of createTwilioFunction for easier testing and isolation.
