+++ packages/create-twilio-function/src/create-twilio-function.js
@@ -1,25 +1,4 @@
-const inquirer = require('inquirer');
-
-async function promptForAccountDetails(config) {
-  const questions = [];
-  if (!config.accountSid) {
-    questions.push({
-      type: 'input',
-      name: 'accountSid',
-      message: 'Twilio Account SID',
-      validate: input =>
-        input.startsWith('AC') ? true : 'An Account SID starts with "AC".'
-    });
-  }
-  if (!config.authToken) {
-    questions.push({
-      type: 'password',
-      name: 'authToken',
-      message: 'Twilio auth token'
-    });
-  }
-  return await inquirer.prompt(questions);
-}
+const { promptForAccountDetails } = require('./create-twilio-function/prompt');
 
 async function createTwilioFunction(config) {
   const accountDetails = await promptForAccountDetails(config);
